+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Removed node_type option for wazuh-certs-tool.sh. ([#268](https://github.com/wazuh/wazuh-installation-assistant/pull/268))
 - Changed VERSION file to the new standard format. ([#245](https://github.com/wazuh/wazuh-installation-assistant/pull/245))
 - Check available space before the installation with the installation assistant ([#56](https://github.com/wazuh/wazuh-installation-assistant/pull/56))
 - Dependencies installation is reworked in Installation assistant. ([#90](https://github.com/wazuh/wazuh-installation-assistant/pull/90))